Extract Event and AssertionError to toplevel decls
