+++ src/commonMain/kotlin/app/cash/turbine/TurbineAssertionError.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2022 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.turbine
+
+/**
+ * A plain [AssertionError] working around three bugs:
+ *
+ *  1. No two-arg constructor in common (https://youtrack.jetbrains.com/issue/KT-40728).
+ *  2. No two-arg constructor in Java 6.
+ *  3. Public exceptions with public constructors have referential equality broken by coroutines.
+ */
+internal class TurbineAssertionError(
+  message: String,
+  override val cause: Throwable?
+) : AssertionError(message)
